# dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserSignUpResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"preferCategories": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "선호 카테고리 이름 배열"
					}
				},
				"required": [
					"email",
					"name",
					"preferCategories"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserSignUpResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserSignUpResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponse": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false
					},
					"error": {
						"properties": {
							"data": {
								"nullable": true
							},
							"message": {
								"type": "string"
							},
							"errorCode": {
								"type": "string"
							}
						},
						"required": [
							"data",
							"message",
							"errorCode"
						],
						"type": "object"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserSignUpRequest": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일 (로그인 시 사용)"
					},
					"password": {
						"type": "string",
						"description": "유저 비밀번호"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"nickname": {
						"type": "string",
						"description": "유저 닉네임 (선택)"
					},
					"gender": {
						"type": "string",
						"description": "성별 (예: 남성, 여성)"
					},
					"birthDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						],
						"description": "생년월일 (예: 2000-01-01)"
					},
					"address": {
						"type": "string",
						"description": "주소 (선택)"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소 (선택)"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호 (예: 010-1234-5678)"
					},
					"preferences": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"description": "선호 카테고리 ID 배열 (예: [1, 2])"
					}
				},
				"required": [
					"email",
					"password",
					"name",
					"gender",
					"birthDate",
					"phoneNumber",
					"preferences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateResponse": {
				"properties": {
					"email": {
						"type": "string",
						"description": "유저 이메일"
					},
					"name": {
						"type": "string",
						"description": "유저 이름"
					},
					"nickname": {
						"type": "string",
						"description": "유저 닉네임"
					},
					"gender": {
						"type": "string",
						"description": "성별"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time",
						"description": "생년월일"
					},
					"address": {
						"type": "string",
						"description": "주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호"
					}
				},
				"required": [
					"email",
					"name",
					"nickname",
					"gender",
					"birthDate",
					"address",
					"detailAddress",
					"phoneNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserUpdateResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserUpdateResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserUpdateRequest": {
				"properties": {
					"nickname": {
						"type": "string",
						"description": "유저 닉네임"
					},
					"gender": {
						"type": "string",
						"description": "성별 (예: 남성, 여성)"
					},
					"birthDate": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							}
						],
						"description": "생년월일 (예: 2000-01-01)"
					},
					"address": {
						"type": "string",
						"description": "주소"
					},
					"detailAddress": {
						"type": "string",
						"description": "상세 주소"
					},
					"phoneNumber": {
						"type": "string",
						"description": "전화번호 (예: 010-1234-5678)"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "사용자 미션 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"missionName": {
						"type": "string",
						"nullable": true,
						"description": "미션 이름"
					},
					"status": {
						"type": "string",
						"description": "미션 상태 (예: ONGOING, COMPLETE)"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"description": "미션 시작일"
					},
					"completedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "미션 완료일"
					},
					"progressCount": {
						"type": "number",
						"format": "double",
						"description": "미션 진행 횟수"
					},
					"rewardPoint": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"endDate": {
						"type": "string",
						"format": "date-time",
						"nullable": true,
						"description": "미션 종료일"
					}
				},
				"required": [
					"id",
					"userId",
					"missionId",
					"missionName",
					"status",
					"startedAt",
					"completedAt",
					"progressCount",
					"rewardPoint",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionItem_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserMissionItem"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserMissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserMissionItem"
						},
						"type": "array",
						"description": "사용자 미션 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true,
								"description": "다음 페이지 커서 (없으면 null)"
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 정보"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserMissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/UserMissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					},
					"category": {
						"type": "string",
						"description": "가게 카테고리"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "생성일"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "수정일"
					}
				},
				"required": [
					"id",
					"regionId",
					"name",
					"address",
					"category",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_StoreResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/StoreResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreAddRequest": {
				"properties": {
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "지역 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					},
					"category": {
						"type": "string",
						"description": "가게 카테고리"
					}
				},
				"required": [
					"regionId",
					"name",
					"address",
					"category"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"userId": {
						"type": "number",
						"format": "double",
						"description": "유저 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점 (1~5)"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "생성일"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "수정일"
					}
				},
				"required": [
					"id",
					"userId",
					"storeId",
					"content",
					"rating",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewItem_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewItem"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_ReviewAddBody.Exclude_keyofReviewAddBody.userId__": {
				"properties": {
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점 (1~5)"
					}
				},
				"required": [
					"content",
					"rating"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_ReviewAddBody.userId_": {
				"$ref": "#/components/schemas/Pick_ReviewAddBody.Exclude_keyofReviewAddBody.userId__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"ReviewListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewItem"
						},
						"type": "array",
						"description": "리뷰 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true,
								"description": "다음 페이지 커서 (없으면 null)"
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 정보"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ReviewListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/ReviewListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"name": {
						"type": "string",
						"description": "미션 이름"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "보상 포인트"
					},
					"condition": {
						"type": "string",
						"description": "미션 조건"
					},
					"conditionValue": {
						"type": "number",
						"format": "double",
						"description": "조건 값"
					},
					"startDate": {
						"type": "string",
						"format": "date-time",
						"description": "미션 시작일"
					},
					"endDate": {
						"type": "string",
						"format": "date-time",
						"description": "미션 종료일"
					},
					"missionSpec": {
						"type": "string",
						"description": "미션 상세 설명"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "생성일"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "수정일"
					}
				},
				"required": [
					"id",
					"storeId",
					"name",
					"reward",
					"condition",
					"conditionValue",
					"startDate",
					"endDate",
					"missionSpec",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionItem_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionItem"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionAddBody": {
				"properties": {
					"name": {
						"type": "string",
						"description": "미션 이름"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "보상 포인트"
					},
					"condition": {
						"type": "string",
						"description": "미션 조건"
					},
					"conditionValue": {
						"type": "number",
						"format": "double",
						"description": "조건 값"
					},
					"startDate": {
						"type": "string",
						"description": "미션 시작일 (예: 2024-01-01)"
					},
					"endDate": {
						"type": "string",
						"description": "미션 종료일 (예: 2024-12-31)"
					},
					"missionSpec": {
						"type": "string",
						"description": "미션 상세 설명"
					}
				},
				"required": [
					"name",
					"reward",
					"condition",
					"conditionValue",
					"startDate",
					"endDate",
					"missionSpec"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/MissionItem"
						},
						"type": "array",
						"description": "미션 목록"
					},
					"pagination": {
						"properties": {
							"cursor": {
								"type": "number",
								"format": "double",
								"nullable": true,
								"description": "다음 페이지 커서 (없으면 null)"
							}
						},
						"required": [
							"cursor"
						],
						"type": "object",
						"description": "페이지네이션 정보"
					}
				},
				"required": [
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_MissionListResponse_": {
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"data": {
						"$ref": "#/components/schemas/MissionListResponse"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "umc",
		"version": "1.0.0",
		"description": "UMC project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users/signup": {
			"post": {
				"operationId": "HandleUserSignUp",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserSignUpResponse_"
								}
							}
						}
					},
					"409": {
						"description": "중복된 이메일 에러",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "회원가입 API",
				"summary": "회원가입을 처리하는 엔드포인트입니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserSignUpRequest"
							}
						}
					}
				}
			}
		},
		"/users/me": {
			"patch": {
				"operationId": "HandleUserUpdate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserUpdateResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 사용자",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "사용자 정보 수정 API",
				"summary": "로그인한 사용자의 정보를 수정합니다.",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserUpdateRequest"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/missions/{missionId}/challenge": {
			"post": {
				"operationId": "HandleChallengeMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionItem_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 유저 또는 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 도전 API",
				"summary": "유저가 특정 미션에 도전합니다.",
				"tags": [
					"UserMission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/me/missions": {
			"get": {
				"operationId": "HandleGetUserMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 유저",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "유저별 미션 목록 조회 API",
				"summary": "유저의 미션 목록을 조회합니다.",
				"tags": [
					"UserMission"
				],
				"security": [],
				"parameters": [
					{
						"description": "미션 상태 (예: ONGOING, COMPLETE)",
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"default": "ONGOING",
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/missions/{userMissionId}/complete": {
			"post": {
				"operationId": "HandleCompleteUserMission",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserMissionItem_"
								}
							}
						}
					},
					"400": {
						"description": "미션 상태 오류",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 유저 미션",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 완료 처리 API",
				"summary": "진행 중인 미션을 완료 처리합니다.",
				"tags": [
					"UserMission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "userMissionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores": {
			"post": {
				"operationId": "HandleStoreAdd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_StoreResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 지역",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 등록된 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "가게 추가 API",
				"summary": "새로운 가게를 등록합니다.",
				"tags": [
					"Store"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/StoreAddRequest"
							}
						}
					}
				}
			}
		},
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "HandleReviewAdd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewItem_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게 또는 유저",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "리뷰 추가 API",
				"summary": "가게에 리뷰를 등록합니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_ReviewAddBody.userId_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListStoreReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "가게 리뷰 목록 조회 API",
				"summary": "특정 가게의 리뷰 목록을 조회합니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/users/me/reviews": {
			"get": {
				"operationId": "HandleListUserReviews",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ReviewListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 유저",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "유저 리뷰 목록 조회 API",
				"summary": "특정 유저의 리뷰 목록을 조회합니다.",
				"tags": [
					"Review"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/stores/{storeId}/missions": {
			"post": {
				"operationId": "HandleMissionAdd",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionItem_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "미션 추가 API",
				"summary": "가게에 새로운 미션을 등록합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MissionAddBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "HandleListStoreMissions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_MissionListResponse_"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"description": "가게 미션 목록 조회 API",
				"summary": "특정 가게의 미션 목록을 조회합니다.",
				"tags": [
					"Mission"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}